+++ settings/index.js
@@ -4,6 +4,7 @@ import Settings from '@folio/stripes-components/lib/Settings';
 import Configuration from './Configuration';
 import TestHotkeys from './TestHotkeys';
 import TestPlugin from './TestPlugin';
+import Token from './Token';
 
 const pages = [
   {
@@ -20,6 +21,11 @@ const pages = [
     route: 'plugin',
     label: 'Plugin Test',
     component: TestPlugin,
+  },
+  {
+    route: 'token',
+    label: 'Set Token',
+    component: Token,
   },
 ];